+++ Tests/SentryTests/SentryTests-Bridging-Header.h
@@ -196,8 +196,8 @@
 #import "SentrySpanOperations.h"
 #import "SentryTimeToDisplayTracker.h"
 #import "SentryTracerConfiguration.h"
-#import "TestSentryViewHierarchy.h"
 #import "TestSentrySingleExecution.h"
+#import "TestSentryViewHierarchy.h"
 #if SENTRY_HAS_UIKIT
 #    import "MockUIScene.h"
 #    import "SentryUIEventTracker.h"
